+++ packages/SwingSet/test/basedir-promises-2/bootstrap.js
@@ -1,5 +1,5 @@
 import harden from '@agoric/harden';
-import { makePromise } from '@agoric/make-promise';
+import { producePromise } from '@agoric/produce-promise';
 
 console.log(`loading bootstrap`);
 
@@ -8,7 +8,7 @@ function build(E, log) {
     bootstrap(argv, vats) {
       const mode = argv[0];
       if (mode === 'harden-promise-1') {
-        const { promise: p1 } = makePromise();
+        const { promise: p1 } = producePromise();
         harden(p1);
         const allP = [];
         // in bug #95, this first call returns a (correctly) frozen Promise,