+++ packages/produce-promise/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.4](https://github.com/Agoric/agoric-sdk/compare/@agoric/make-promise@0.0.3-alpha.0...@agoric/produce-promise@0.0.3) (2020-04-06)
+
+renamed package from @agoric/make-promise
+
+Renamed from { p, res, reject } to { promise, resolve, reject }
+
 ## [0.0.3](https://github.com/Agoric/agoric-sdk/compare/@agoric/make-promise@0.0.3-alpha.0...@agoric/make-promise@0.0.3) (2020-04-02)
 
 **Note:** Version bump only for package @agoric/make-promise